refactor mouse events out of useOcifEditor to useMouseEvents and actions
